requirements.txt
fastapi
uvicorn[standard]
watchdog
pydantic
python-dotenv
faiss-cpu
sentence-transformers
langchain
httpx
pytest